+++ ProjectTRM/Objects/Character/Enemy/Hertic/Heretic.h
@@ -113,7 +113,7 @@ public:
 	/// HPŠÇ—ˆ—
 	/// </summary>
 	/// <param name="hit_object">ƒ_ƒ[ƒW</param>
-	virtual void HPControl(int Damage);
+	virtual void HPControl(float Damage);
 
 	bool GetDead();
