+++ ProjectTRM/Objects/Character/Player/Oracle/Oracle.h
@@ -58,7 +58,7 @@ public:
 	/// HPŠÇ—ˆ—
 	/// </summary>
 	/// <param name="hit_object">ƒ_ƒ[ƒW</param>
-	virtual void HPControl(int Damage);
+	virtual void HPControl(float Damage);
 
 private:
 	/// <summary>
